refactor: moving the shell helpers into the shell target dir
