+++ src/targets/shell/wget.js
@@ -11,7 +11,7 @@
 'use strict'
 
 var util = require('util')
-var helpers = require('../../helpers/shell')
+var helpers = require('./helpers')
 var CodeBuilder = require('../../helpers/code-builder')
 
 module.exports = function (source, options) {
